fix 2.0-beta bugs
