+++ configs/animeganv2.yaml
@@ -10,7 +10,7 @@ model:
   gan_criterion:
     name: GANLoss
     gan_mode: lsgan
-  pretrain_ckpt: output_dir/AnimeGANV2PreTrainModel-2020-11-29-17-02/epoch_2_checkpoint.pdparams
+  pretrain_ckpt: ~ #output_dir/AnimeGANV2PreTrainModel-2020-11-29-17-02/epoch_2_checkpoint.pdparams
   g_adv_weight: 300.
   d_adv_weight: 300.
   con_weight: 1.5
@@ -47,21 +47,39 @@ dataset:
   test:
     name: SingleDataset
     dataroot: data/animedataset/test/HR_photo
-    max_dataset_size: inf
-    direction: BtoA
-    input_nc: 3
-    output_nc: 3
-    serial_batches: False
-    pool_size: 50
-    transforms:
-      - name: ResizeToScale
-        size: [256, 256]
-        scale: 32
-        interpolation: bilinear
-      - name: Transpose
-      - name: Normalize
-        mean: [127.5, 127.5, 127.5]
-        std: [127.5, 127.5, 127.5]
+    preprocess:
+      - name: LoadImageFromFile
+        key: A
+      - name: Transforms
+        input_keys: [A]
+        pipeline:
+          - name: ResizeToScale
+            size: [256, 256]
+            scale: 32
+            interpolation: bilinear
+          - name: Transpose
+          - name: Normalize
+            mean: [127.5, 127.5, 127.5]
+            std: [127.5, 127.5, 127.5]
+            keys: [image, image]
+  # test:
+  #   name: SingleDataset
+  #   dataroot: data/animedataset/test/HR_photo
+  #   max_dataset_size: inf
+  #   direction: BtoA
+  #   input_nc: 3
+  #   output_nc: 3
+  #   serial_batches: False
+  #   pool_size: 50
+  #   transforms:
+  #     - name: ResizeToScale
+  #       size: [256, 256]
+  #       scale: 32
+  #       interpolation: bilinear
+  #     - name: Transpose
+  #     - name: Normalize
+  #       mean: [127.5, 127.5, 127.5]
+  #       std: [127.5, 127.5, 127.5]
 
 lr_scheduler:
   name: LinearDecay
