--- ppgan/datasets/transforms/__init__.py
@@ -1,1 +1,0 @@
-from .transforms import ResizeToScale, PairedRandomCrop, PairedRandomHorizontalFlip, Add
