+++ ppgan/models/dc_gan_model.py
@@ -56,8 +56,9 @@ class DCGANModel(BaseModel):
             input (dict): include the data itself and its metadata information.
         """
         # get 1-channel gray image, or 3-channel color image
-        self.real = paddle.to_tensor(input['A'])
-        self.image_paths = input['A_path']
+        self.real = paddle.to_tensor(input['img'])
+        if 'img_path' in input:
+            self.image_paths = input['A_path']
 
     def forward(self):
         """Run forward pass; called by both functions <train_iter> and <test_iter>."""
